feat(Voting): add reference field error message
